requirements.txt
pika>=1.3.0
requests>=2.31.0
openai-whisper>=20231117
numpy>=1.24.0
torch>=2.0.0